move checkResponseOK to api.ts
